Refactor settings classes
They now inherit from a settings abstract class
