+++ src/Settings/HomesteadSettings.php
@@ -2,15 +2,32 @@
 
 namespace Laravel\Homestead\Settings;
 
-interface HomesteadSettings
+abstract class HomesteadSettings
 {
+    /**
+     * Settings attributes.
+     *
+     * @var array
+     */
+    protected $attributes;
+
+    /**
+     * JsonSettings constructor.
+     *
+     * @param  array  $attributes
+     */
+    public function __construct($attributes)
+    {
+        $this->attributes = $attributes;
+    }
+
     /**
      * Create an instance from a file.
      *
      * @param  string  $filename
      * @return static
      */
-    public static function fromFile($filename);
+    abstract public static function fromFile($filename);
 
     /**
      * Save the homestead settings.
@@ -18,15 +35,22 @@ interface HomesteadSettings
      * @param  string  $filename
      * @return void
      */
-    public function save($filename);
+    abstract public function save($filename);
 
     /**
      * Update the homestead settings.
      *
      * @param  array  $attributes
-     * @return void
+     * @return static
      */
-    public function update($attributes);
+    public function update($attributes)
+    {
+        $this->attributes = array_merge($this->attributes, array_filter($attributes, function ($attribute) {
+            return ! is_null($attribute);
+        }));
+
+        return $this;
+    }
 
     /**
      * Update the virtual machine's name.
@@ -34,7 +58,12 @@ interface HomesteadSettings
      * @param  string  $name
      * @return static
      */
-    public function updateName($name);
+    public function updateName($name)
+    {
+        $this->update(['name' => $name]);
+
+        return $this;
+    }
 
     /**
      * Update the virtual machine's hostname.
@@ -42,7 +71,12 @@ interface HomesteadSettings
      * @param  string  $hostname
      * @return static
      */
-    public function updateHostname($hostname);
+    public function updateHostname($hostname)
+    {
+        $this->update(['hostname' => $hostname]);
+
+        return $this;
+    }
 
     /**
      * Update the virtual machine's IP address.
@@ -50,16 +84,37 @@ interface HomesteadSettings
      * @param  string  $ip
      * @return static
      */
-    public function updateIpAddress($ip);
+    public function updateIpAddress($ip)
+    {
+        $this->update(['ip' => $ip]);
+
+        return $this;
+    }
 
     /**
      * Configure the nginx sites.
      *
      * @param  string  $projectName
-     * @param  string  $slugifiedProjectName
      * @return static
      */
-    public function configureSites($projectName, $slugifiedProjectName);
+    public function configureSites($projectName, $slugifiedProjectName)
+    {
+        $site = [
+            'map' => "{$projectName}.app",
+            'to' => "/home/vagrant/Code/{$slugifiedProjectName}/public",
+        ];
+
+        if (isset($this->attributes['sites']) &&
+            ! empty($this->attributes['sites']) &&
+            isset($this->attributes['sites'][0]['schedule'])
+        ) {
+            $site['schedule'] = $this->attributes['sites'][0]['schedule'];
+        }
+
+        $this->update(['sites' => [$site]]);
+
+        return $this;
+    }
 
     /**
      * Configure the shared folders.
@@ -68,12 +123,25 @@ interface HomesteadSettings
      * @param  string  $slugifiedProjectName
      * @return static
      */
-    public function configureSharedFolders($projectPath, $slugifiedProjectName);
+    public function configureSharedFolders($projectPath, $slugifiedProjectName)
+    {
+        $folder = [
+            'map' => $projectPath,
+            'to' => "/home/vagrant/Code/{$slugifiedProjectName}",
+        ];
+
+        $this->update(['folders' => [$folder]]);
+
+        return $this;
+    }
 
     /**
      * Convert the homestead settings to an array.
      *
      * @return array
      */
-    public function toArray();
+    public function toArray()
+    {
+        return $this->attributes;
+    }
 }
